requirements.txt
Pillow
requests
smbus2
RPi.GPIO
RPi.bme280